Move Array to shared partition
